Change error message and check url id

+++ backend/controllers/product.controller.js
@@ -53,11 +53,17 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
   const { id } = req.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return res
+      .status(500)
+      .json({ success: false, message: "Invalid Product Id" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     res.status(200).json({ success: true, message: "Product Deleted" });
   } catch (err) {
     console.error("Error in Deleting products" + err.message);
-    res.status(404).json({ success: false, message: "Product not Found" });
+    res.status(500).json({ success: false, message: "Server Error" });
   }
 };
